Overhaul the testing framework to enable testing of keyword arguments & default arguments.
The new testing framework also provides stronger guarantees about the correctness of the test-cases themselves:  Guarantees can now be checked at module loading-time, when each test-case is constructed, rather than when each test-case is run.

I've added 3 new testing-framework classes in module `utils_for_tests.py`:

- class `TestCase`
- class `ExpectedReturn`
- class `ExpectedException`

Class `TestCase` replaces namedtuple `Test`, which was constructed with 7 arguments that could not be checked for type-correctness until the `Test` instance was examined (because a namedtuple has no `__init__` method).  (With that many arguments, there's an easy risk of missing an argument or duplicating an argument.)

Now class `TestCase` is able to validate its argument-types in its constructor (`__init__` method), when the `TestCase` is instantiated at module-loading time.

Class `ExpectedReturn` provides new functionality, to enable testing of keyword arguments & default arguments (supported by new code in function `_run_test`).  It allows us to check for return-values that we passed into the test-function ourselves (possibly randomly-generated) as either positional or keyword arguments, and/or for hard-coded default arguments if we don't pass in arguments ourselves.

Class `ExpectedException` refactors-out 3 arguments from the constructor of class `TestCase` (the `type()`, `repr()`, and `str()`, of a raised exception).  This provides 2 benefits:

1. Fewer arguments to the constructor of `TestCase`: Now, 3 of them are instead grouped into the constructor of `ExpectedException` (where they can again be type-checked by the `ExpectedException` `__init__` method).

2. Having distinct classes `ExpectedReturn` & `ExpectedException` enables us to verify that **only** a return-value **or** a raised exception is expected in any given `TestCase` (in its new attribute `TestCase.expected`) --- not both!

This increases semantic correctness, but also removes some redundant arguments: Previously, if a return-value was expected in a test-case, the 3 exception-specific arguments were still specified as `None, None, None`.

As a result, class `TestCase` now takes only 5 arguments, all of which may be type-checked in the `__init__` method, and none of which may be `None`.

And finally, after I implemented this new testing-framework, I was able to add some test-cases that test keyword arguments & default arguments!