Rename check class `isTypeSequence` to `isTypeOfSequence`.
Also added a long comment in module `argcheck/impl.py` to document the different behaviour of `typing.Sequence` vs `collections.abc.Sequence` on different versions of Python3, and to explain why this code has made the choices it has.  For the benefit of future generations.

Also added a medium-length comment in method `isTypeOfSequence.is_valid` to explain the "why" of the current implementation of this method.